test: fold source route coverage into squad page suite
